Fixed folder structure for v2x module
There was one folder `com.foursoft.harness.vec.v2x` instead
of a directory with further directories (`com/foursoft/harness/vec/v2x`).
