+++ v2x/src/main/java/com/foursoft/harness/vec/v2x/predicates/VecPredicates.java
