+++ v2x/src/main/java/com/foursoft/harness/vec/v2x/visitor/StrictBaseVisitor.java
